requirements.txt
absl-py==0.7.1
astor==0.7.1
gast==0.2.2
grpcio==1.20.1
Markdown==3.1
numpy==1.14.5
opencv-python==4.1.0.25
Pillow==6.0.0
protobuf==3.6.0
scipy==1.2.1
six==1.12.0
tensorboard==1.8.0
tensorflow-gpu==1.8.0
termcolor==1.1.0
Werkzeug==0.15.2
